Small cleanup refactor in UI and moved methods to UIDocumentExtensions
