+++ sources/engine/Stride.UI/UISystem.Picking.cs
@@ -4,11 +4,8 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Stride.Core;
-using Stride.Core.Collections;
 using Stride.Core.Diagnostics;
 using Stride.Core.Mathematics;
-using Stride.Engine;
 using Stride.Games;
 using Stride.Graphics;
 using Stride.Input;
@@ -22,9 +19,8 @@ namespace Stride.UI
         // object to avoid allocation at each element leave event
         private readonly HashSet<UIElement> newlySelectedElementParents = new HashSet<UIElement>();
         private readonly List<PointerEvent> compactedPointerEvents = new List<PointerEvent>();
-        private readonly List<RenderUIPage> previousRenderObjects = new List<RenderUIPage>();
         
-        private TrackingHashSet<UIDocument> documents = new TrackingHashSet<UIDocument>(); 
+        private readonly HashSet<UIDocument> documents = new HashSet<UIDocument>(); 
         
         /// <summary>
         /// Represents the UI-element that's currently under the mouse cursor.
@@ -32,16 +28,6 @@ namespace Stride.UI
         /// Last processed element_state / ?UIComponent? with a valid element will be used.
         /// </summary>
         public UIElement UIElementUnderMouseCursor { get; internal set; }
-
-        public bool AddDocument(UIDocument document)
-        {
-            return documents.Add(document);
-        }
-        
-        public bool RemoveDocument(UIDocument document)
-        {
-            return documents.Add(document);
-        }
 
         private partial void Pick(GameTime gameTime)
         {
@@ -61,7 +47,7 @@ namespace Stride.UI
             
                 foreach (var uiDocument in documents)
                 {
-                    Matrix worldViewProjection = GetWorldViewProjection(uiDocument, cameraSlot.Camera, renderTarget);
+                    Matrix worldViewProjection = uiDocument.GetWorldViewProjection(cameraSlot.Camera, renderTarget);
                     
                     // Check if the current UI component is being picked based on the current ViewParameters (used to draw this element)
                     using (Profiler.Begin(UIProfilerKeys.TouchEventsUpdate))
@@ -89,8 +75,8 @@ namespace Stride.UI
             var inverseZViewProj = worldViewProj;
             inverseZViewProj.Row3 = -inverseZViewProj.Row3;
 
-            elementUnderPointer = UpdateMouseOver(ref viewport, ref inverseZViewProj, uiDocument);
-            UpdateTouchEvents(ref viewport, ref inverseZViewProj, uiDocument, gameTime);
+            elementUnderPointer = UpdateMouseOver(uiDocument, ref viewport, ref inverseZViewProj);
+            UpdateTouchEvents(uiDocument, ref viewport, ref inverseZViewProj, gameTime);
         }
 
         private void PickingPrepare()
@@ -191,7 +177,7 @@ namespace Stride.UI
             return uiRay;
         }
 
-        private void UpdateTouchEvents(ref Viewport viewport, ref Matrix worldViewProj, UIDocument uiDocument, GameTime gameTime)
+        private void UpdateTouchEvents(UIDocument uiDocument, ref Viewport viewport, ref Matrix worldViewProj, GameTime gameTime)
         {
             var rootElement = uiDocument.Page.RootElement;
             var intersectionPoint = Vector3.Zero;
@@ -289,7 +275,7 @@ namespace Stride.UI
             }
         }
 
-        private UIElement UpdateMouseOver(ref Viewport viewport, ref Matrix worldViewProj, UIDocument uiDocument)
+        private UIElement UpdateMouseOver(UIDocument uiDocument, ref Viewport viewport, ref Matrix worldViewProj)
         {
             if (input == null || !input.HasMouse)
                 return null;
@@ -501,122 +487,31 @@ namespace Stride.UI
                 PerformRecursiveHitTest(child, ref ray, ref worldViewProj, results);
         }
 
-        public static Matrix GetWorldViewProjection(UIDocument uiDocument,  CameraComponent camera, Texture renderTarget)
+        /// <summary>
+        /// Adds the specified <see cref="UIDocument"/> to the <see cref="UISystem"/> to receive input events.
+        /// </summary>
+        /// <param name="document">The document to add.</param>
+        /// <returns><c>true</c> if the document is added to the <see cref="UISystem"/>; <c>false</c> if the document is already present.</returns>
+        /// <remarks>A <see cref="RenderUIDocument"/> render object needs to be added to the <see cref="RenderSystem"/> for the UI to be rendered.</remarks>
+        public bool AddDocument(UIDocument document)
         {
-            Matrix worldViewProjection = Matrix.Identity;
-                    
-            // calculate the size of the virtual resolution depending on target size (UI canvas)
-            var virtualResolution = uiDocument.Resolution;
-
-            if (uiDocument.IsFullScreen)
-            {
-                //var targetSize = viewportSize;
-                var targetSize = new Vector2(renderTarget.Width, renderTarget.Height);
-
-                // update the virtual resolution of the renderer
-                if (uiDocument.ResolutionStretch == ResolutionStretch.FixedWidthAdaptableHeight)
-                    virtualResolution.Y = virtualResolution.X * targetSize.Y / targetSize.X;
-                if (uiDocument.ResolutionStretch == ResolutionStretch.FixedHeightAdaptableWidth)
-                    virtualResolution.X = virtualResolution.Y * targetSize.X / targetSize.Y;
-
-                worldViewProjection = GetWorldViewProjection(uiDocument, virtualResolution);
-            }
-            else
-            {
-                //var cameraComponent = renderContext.Tags.Get(CameraComponentRendererExtensions.Current);
-                if (camera != null)
-                    worldViewProjection = GetWorldViewProjection(uiDocument, camera);
-            }
-
-            return worldViewProjection;
+            return documents.Add(document);
         }
         
-        private static Matrix GetWorldViewProjection(UIDocument uiDocument, CameraComponent camera)
+        /// <summary>
+        /// Removes the specified <see cref="UIDocument"/> from the <see cref="UISystem"/>.
+        /// </summary>
+        /// <param name="document">The document to remove.</param>
+        /// <returns>
+        /// <c>true</c> if the document was successfully removed; otherwise <c>false</c>.
+        /// This method also returns <c>false</c> the document was not present.
+        /// </returns>
+        /// <remarks>A <see cref="RenderUIDocument"/> render object needs to be removed from the <see cref="RenderSystem"/> for the UI to stop being rendered.</remarks>
+        public bool RemoveDocument(UIDocument document)
         {
-            var frustumHeight = 2 * MathF.Tan(MathUtil.DegreesToRadians(camera.VerticalFieldOfView) / 2);
-
-            var worldMatrix = uiDocument.WorldMatrix;
-
-            // rotate the UI element perpendicular to the camera view vector, if billboard is activated
-            if (uiDocument.IsFullScreen)
-            {
-                worldMatrix = Matrix.Identity;
-            }
-            else
-            {
-                Matrix viewInverse;
-                Matrix.Invert(ref camera.ViewMatrix, out viewInverse);
-                var forwardVector = viewInverse.Forward;
-
-                if (uiDocument.IsBillboard)
-                {
-                    var viewInverseRow1 = viewInverse.Row1;
-                    var viewInverseRow2 = viewInverse.Row2;
-
-                    // remove scale of the camera
-                    viewInverseRow1 /= viewInverseRow1.XYZ().Length();
-                    viewInverseRow2 /= viewInverseRow2.XYZ().Length();
-
-                    // set the scale of the object
-                    viewInverseRow1 *= worldMatrix.Row1.XYZ().Length();
-                    viewInverseRow2 *= worldMatrix.Row2.XYZ().Length();
-
-                    // set the adjusted world matrix
-                    worldMatrix.Row1 = viewInverseRow1;
-                    worldMatrix.Row2 = viewInverseRow2;
-                    worldMatrix.Row3 = viewInverse.Row3;
-                }
-
-                if (uiDocument.IsFixedSize)
-                {
-                    forwardVector.Normalize();
-                    var distVec = (worldMatrix.TranslationVector - viewInverse.TranslationVector);
-                    float distScalar;
-                    Vector3.Dot(ref forwardVector, ref distVec, out distScalar);
-                    distScalar = Math.Abs(distScalar);
-
-                    var worldScale = frustumHeight * distScalar * UIComponent.FixedSizeVerticalUnit; // FrustumHeight already is 2*Tan(FOV/2)
-
-                    worldMatrix.Row1 *= worldScale;
-                    worldMatrix.Row2 *= worldScale;
-                    worldMatrix.Row3 *= worldScale;
-                }
-
-                // If the UI component is not drawn fullscreen it should be drawn as a quad with world sizes corresponding to its actual size
-                worldMatrix = Matrix.Scaling(uiDocument.Size / uiDocument.Resolution) * worldMatrix;
-            }
-
-            // Rotation of Pi along 0x to go from UI space to world space
-            worldMatrix.Row2 = -worldMatrix.Row2;
-            worldMatrix.Row3 = -worldMatrix.Row3;
-
-            Matrix worldViewMatrix;
-            Matrix worldViewProjectionMatrix;
-            Matrix.Multiply(ref worldMatrix, ref camera.ViewMatrix, out worldViewMatrix);
-            Matrix.Multiply(ref worldViewMatrix, ref camera.ProjectionMatrix, out worldViewProjectionMatrix);
-
-            return worldViewProjectionMatrix;
+            return documents.Remove(document);
         }
-
-        private static Matrix GetWorldViewProjection(UIDocument uiDocument, Vector3 virtualResolution)
-        {
-            var nearPlane = virtualResolution.Z / 2;
-            var farPlane = nearPlane + virtualResolution.Z;
-            var zOffset = nearPlane + virtualResolution.Z / 2;
-            var aspectRatio = virtualResolution.X / virtualResolution.Y;
-            var verticalFov = MathF.Atan2(virtualResolution.Y / 2, zOffset) * 2;
-
-            var cameraComponent = new CameraComponent(nearPlane, farPlane)
-            {
-                UseCustomAspectRatio = true,
-                AspectRatio = aspectRatio,
-                VerticalFieldOfView = MathUtil.RadiansToDegrees(verticalFov),
-                ViewMatrix = Matrix.LookAtRH(new Vector3(0, 0, zOffset), Vector3.Zero, Vector3.UnitY),
-                ProjectionMatrix = Matrix.PerspectiveFovRH(verticalFov, aspectRatio, nearPlane, farPlane),
-            };
-
-            return GetWorldViewProjection(uiDocument, cameraComponent);
-        }
+        
 
         /// <summary>
         /// Represents the result of a hit test on the UI.
@@ -641,16 +536,6 @@ namespace Stride.UI
             /// Point of intersection between the ray and the hit element.
             /// </summary>
             public Vector3 IntersectionPoint { get; }
-        }
-    }
-
-    public class UIElementInputState
-    {
-        
-
-        public UIElementInputState()
-        {
-            
         }
     }
 }
